[citest skip] Add changelog_to_tag.yml to .github/workflows
Description:
When a new changelog section is added to CHANGELOG.md and pushed,
changelog_to_tag.yml is triggered, which generates a new tag and
a new release.

Example of CHANGELOG.md changes:
  [9.9.9] - 2022-12-31
  --------------------

  ### New features

  - New feature A

  ### Bug fixes

  - Bug fix B

Using this example, when the commit on CHANGELOG.md is pushed, a
new tag "9.9.9" is added and Version 9.9.9 is released in github.
If tag "9.9.9" already exists, the CHANGELOG.md push fails.

NOTE: Bootloader role is not tagged with any version yet. When it
is, please add CHANGELOG.md to the bootloader top directory having
the following contents, where #.#.# is the initial version and YYYY
-MM-DD is the date adding the CHANGELOG.md. And the first two white
spaces are to be removed when create CHANGELOG.md.

  Changelog
  =========

  [#.#.#] - YYYY-MM-DD
  --------------------

  # Initial Release

Signed-off-by: Noriko Hosoi <nhosoi@redhat.com>