feat: add support for ostree systems
Feature: Add the ability for the role to manage ostree systems.

Reason: It is necessary to be able to manage ostree systems.

Result: The role is able to manage ostree systems.

NOTE That some of the bootloader functionality may not be available
on ostree systems - specifically, rebooting may not work depending on
how the image is built.

Signed-off-by: Rich Megginson <rmeggins@redhat.com>