ci: Fix broken __default_args detection
tests_add_rm clones the current kernel, and then wants to determine
`__default_args` from the original kernel. But that's precisely *not*
the one which is the default one after cloning (as Clone1 is now the new
and current default).

So instead of determining the attributes from the default kernel, select
the original kernel by title.

Fixes https://issues.redhat.com/browse/RHELMISC-11212