fix: boolean values and null values are not allowed
Cause: The user could specify a value like `value: on` or `value: yes` thinking that these would
be converted to the string `"on"` or `"yes"`. But instead, YAML treats these
as [YAML bool type](https://yaml.org/type/bool.html) and writes them as the string `"True"`.

Consequence: The user who is unaware of YAML boolean handling is not able to set
values to `"on"`, `"off"`, or the like.

Fix: The role will reject any value of boolean or `null` type.

Result: Users must quote such YAML boolean type values as strings in order to write
them to the bootloader configuration.

Signed-off-by: Rich Megginson <rmeggins@redhat.com>